Refactor: Rename "gate_map.rs" to "synched_map.rs"
